requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
pydantic==2.7.1
pyyaml
streamlit